+++ app-code/microservices/insurance-policy/insurance-policy-api/Services/Interfaces/IInsurancePolicyService.cs
@@ -34,7 +34,7 @@ namespace Insurance.Policy.Api.Services.Interfaces
         /// </summary>
         /// <returns>The requested record else NULL</returns>
         /// <param name="id">Identifier.</param>
-        InsurancePolicy FindById(int id);
+        InsurancePolicy FindById(long id);
 
         /// <summary>
         /// Inserts a new record to database storage.
@@ -55,6 +55,6 @@ namespace Insurance.Policy.Api.Services.Interfaces
         /// </summary>
         /// <returns>The delete.</returns>
         /// <param name="id">Identifier.</param>
-        int Delete(int id);
+        int Delete(long id);
     }
 }
